requirements.txt
coverage
psycopg2-binary
pytest